Added all requested change
