+++ app/src/main/java/org/listenbrainz/android/model/ListenBitmap.kt
@@ -1,0 +1,8 @@
+package org.listenbrainz.android.model
+
+import android.graphics.Bitmap
+
+data class ListenBitmap(
+    val bitmap: Bitmap?=null,
+    val id:String?=""
+)
